refactor project view and embed view to not use shared hoc
