+++ src/views/preview/embed-view.jsx
@@ -1,24 +1,26 @@
 // embed view
 
-const bindAll = require('lodash.bindall');
 const React = require('react');
 const PropTypes = require('prop-types');
+const connect = require('react-redux').connect;
 const injectIntl = require('react-intl').injectIntl;
 
-const ProjectViewHOC = require('./project-view-hoc.jsx');
 const ErrorBoundary = require('../../components/errorboundary/errorboundary.jsx');
 const projectShape = require('./projectshape.jsx').projectShape;
 const NotAvailable = require('../../components/not-available/not-available.jsx');
 const Meta = require('./meta.jsx');
 
+const previewActions = require('../../redux/preview.js');
+
 const GUI = require('scratch-gui');
 const IntlGUI = injectIntl(GUI.default);
+
+const initSentry = require('../../lib/sentry.js');
+initSentry();
 
 class EmbedView extends React.Component {
     constructor (props) {
         super(props);
-        bindAll(this, [
-        ]);
         const pathname = window.location.pathname.toLowerCase();
         const parts = pathname.split('/').filter(Boolean);
         this.state = {
@@ -65,7 +67,26 @@ EmbedView.propTypes = {
     projectNotAvailable: PropTypes.bool
 };
 
-module.exports.View = ProjectViewHOC(EmbedView);
+EmbedView.defaultProps = {
+    assetHost: process.env.ASSET_HOST,
+    projectHost: process.env.PROJECT_HOST
+};
+
+const mapStateToProps = state => ({
+    projectInfo: state.preview.projectInfo,
+    projectNotAvailable: state.preview.projectNotAvailable
+});
+
+const mapDispatchToProps = dispatch => ({
+    getProjectInfo: (id, token) => {
+        dispatch(previewActions.getProjectInfo(id, token));
+    }
+});
+
+module.exports.View = connect(
+    mapStateToProps,
+    mapDispatchToProps
+)(EmbedView);
 
 GUI.setAppElement(document.getElementById('app'));
 module.exports.initGuiState = GUI.initEmbedded;
